requirements.txt
Django==1.9
djangorestframework==3.3.1
MySQL-python==1.2.5
Pillow==3.0.0
wheel==0.24.0
